+++ src/settings/BatchGroupConfigurationSettings/BatchVoucherExportsList.js
@@ -9,7 +9,8 @@ import {
 } from '@folio/stripes/components';
 import { FolioFormattedTime } from '@folio/stripes-acq-components';
 
-import { BATCH_VOUCHER_EXPORT_STATUS_LABEL } from '../../../common/constants';
+import { BATCH_VOUCHER_EXPORT_STATUS_LABEL } from '../../common/constants';
+import { RESULT_COUNT_INCREMENT } from './constants';
 
 const columnMapping = {
   date: <FormattedMessage id="ui-invoice.settings.BatchVoucherExports.date" />,
@@ -32,17 +33,24 @@ const cellFormatters = {
   status: ({ status }) => BATCH_VOUCHER_EXPORT_STATUS_LABEL[status],
 };
 
-export function BatchVoucherExportsList({ batchVoucherExports }) {
+export function BatchVoucherExportsList({ batchVoucherExports, onNeedMoreData, recordsCount }) {
+  if (!batchVoucherExports) return null;
+
   return (
     <MultiColumnList
+      autosize
       columnMapping={columnMapping}
       contentData={batchVoucherExports}
       formatter={cellFormatters}
       id="batch-voucher-exports"
       interactive={false}
+      onHeaderClick={() => {}}
+      onNeedMoreData={onNeedMoreData}
+      pageAmount={RESULT_COUNT_INCREMENT}
+      pagingType="click"
       sortDirection="descending"
       sortOrder="date"
-      totalCount={batchVoucherExports.length}
+      totalCount={recordsCount}
       visibleColumns={visibleColumns}
     />
   );
@@ -50,4 +58,6 @@ export function BatchVoucherExportsList({ batchVoucherExports }) {
 
 BatchVoucherExportsList.propTypes = {
   batchVoucherExports: PropTypes.arrayOf(PropTypes.object),
+  onNeedMoreData: PropTypes.func,
+  recordsCount: PropTypes.number,
 };